Getting the correct year end interest's, and getting the correct year end balance's without deposits. However, not getting the correct year end balances with deposits.

+++ UserInputOutput.h
@@ -6,7 +6,7 @@ struct UserInput
 	double iniInvestAmount = 0.0;
 	double monthlyDeposit = 0.0;
 	double annualIntrestRate = 0.0;
-	unsigned short numOfYears = 0.0;
+	int numOfYears = 0;
 };
 
 class UserInputOutput
